# Conflicts:
#	core/src/labels/label.cpp
#	core/src/labels/spriteLabel.cpp
#	core/src/labels/spriteLabel.h
#	core/src/labels/textLabel.cpp
#	osx/src/main.cpp

+++ core/src/scene/styleParam.cpp
@@ -6,6 +6,7 @@
 #include "util/builders.h" // for cap, join
 #include "util/extrude.h"
 #include "util/geom.h" // for CLAMP
+#include "log.h"
 #include <algorithm>
 #include <map>
 #include <cstring>
@@ -162,7 +163,7 @@ StyleParam::Value StyleParam::parseString(StyleParamKey key, const std::string& 
             return std::numeric_limits<uint32_t>::max();
         }
         if (_value == "true") {
-            return 15; // DEFAULT
+            return uint32_t(15); // DEFAULT
         }
         if (parseInt(_value, textWrap) > 0 && textWrap > 0) {
              return static_cast<uint32_t>(textWrap);
